Round 23 audit: live GoingAway/spectator E2E cells and v2-dialect hardening
Close #190 with two pinned live-server cells (SIGTERM drain -> GoingAway ->
close 4000; spectator join/observe/exit) and land audit round 23: widen
RateLimitInfo limits to u64 so a hostile value cannot break the required
Authenticated frame (breaking), reject v2 RoomJoined frames carrying
v3-only token/ICE metadata (mirrors Reconnected), and document that
set_ready() toggles readiness per the protocol authority.

+++ src/protocol.rs
@@ -403,11 +403,15 @@ pub struct PeerConnectionInfo {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct RateLimitInfo {
     /// Requests allowed per minute.
-    pub per_minute: u32,
+    ///
+    /// The protocol authority leaves these limits unbounded integers, so they
+    /// are `u64`: a hostile or absurd `Authenticated` rate-limit value must
+    /// not be able to break authentication.
+    pub per_minute: u64,
     /// Requests allowed per hour.
-    pub per_hour: u32,
+    pub per_hour: u64,
     /// Requests allowed per day.
-    pub per_day: u32,
+    pub per_day: u64,
 }
 
 /// Describes negotiated protocol capabilities for a specific SDK.
